Updated index.html (textually)+(problem statement intro)
I have changed some referenced links of SNS to actual SNS websites, the bottom copyright text to our intuitive team name and agenda, another reference link for an email alongside phone and address. Edited website sub-headline, changed the initial header (that had a fade-in animation once scrolled) MASSIVELY to Fill the turtle form here (that can be referenced with a working Survey123 form link later). I put the objective of PROBLEM STATEMENT in one of the first <divs> that can have a picture of "turtles on the road" and a working rectangular "FULL BRIEF" link that can be clicked upon to know about the entirety of of problem statement info (like open in another tab/within tab like a webpage). I didn't source the picture yet in html but it's stored in the images folder.